Specifies branches as rooted in origin
When we receive events from GitHub, we're told that the branch is
"refs/heads/[name]". This is IN RELATION TO the remote branch. When we
clone internally, since it's a clone, the branch names are
"origin/[name]".

This translation is made at the "acting layer", keeping the GitHub
module relatively clean of the "origin" tag (except in the default
case).