Records if a rule is still in use when removing dependent tables
When removing a rule from a branch, effective and applicable information in
Cassandra should ONLY be removed if the rule is no longer in use. To
drive this test, we record whether rules are in use in a distinct
Cassandra table.

NOTE: We cannot query the rules table because querying by rule_id
would require multi-paritition reads due to the way the primary and
composite keys are designed.

(See: ALLOW FILTERING docs for Cassandra)